Clean up network endpoint abstractions a bit
Before this we had the notion of a "TransportInterface" and "NetInterface".  A "net interface" allowed for establishing new outbound channels, while a "transport interface" only allowed for responding.

We had corresponding "TransportInterfaceSet" and "NetInterfaceSet", but we needed to keep them in sync because components needed to load one for one or the other even if the distinction was irrelevant.

We later modified ServerNode to install "TransportInterfaceSet" into the environment as a "NetInterfaceSet", but didn't remove the places that performed sync.  This ended up messing things up because it deleted all of the "transports" from the "nets set" which were the same object.  This was prematurely removing transports before they could be closed under some circumstances.

There is only a single place where the differentiation mattered, for BLE peripheral interfaces, and there I believe context means you wouldn't attempt to establish outbound connections.  I went ahead and removed NetInterface and made peripheral interfaces just throw an error in this one case if you attempt an outbound connection.

I also renamed the merged interface to "ConnectionlessTransport" because I think that's what the interface actually describes.  "TransportInterface" is confusing because it implies it supports all types of transports, but it has no notion of a connection.  "NetInterface" is doubly confusing because it's also too broad and we generally use that term in the UNIX "interface/link/adapter" sense.