Process and store Read/Subscribe result chunked
This is another optimization of the "legacy controller" to already bring future convenience to it. This change already process changes to data including storage to happen while data is received already.
This can allow reconnection to devices with bad connection because data might fill up over time. I was able to reconnect to a device that was else never finishing.

The new API will do that automatically, so lets use that already here.

The code is not too nice and optimized because potential backport candidate (depending on the conflicts)