Add infrastructure for aborting
I've messed around with cancellation patterns a bit but figured since the JS world is moving toward AbortController we should give that a go.

It's not the nicest API but does work.  This commit adds some utilities that make use more convenient and deletes some of my competing experiments that we never actually used.