Fix Node Shutdown detection in PairedNode
Seems that the former changes in PairedNode in regards to subscriptions had the side effect that after detecting the shutdown event the subscription was reported as successful and so the reconnect was ignored. In fact yes the node went offline after subscription timeout anyway, but should fix that.

Backport candidate.