Always consider client behaviors "supported"
Generates client behavior in a way that we can identify the class.  Then since we know the class interface is a superset of any implementation of the relevant cluster, always return the behavior implementation for the client.

This allows for convenient "mostly typed" access to behaviors regardless of underlying class or enabled cluster features.