Prepare NPM trusted publishing
This PR adjusts the GHA we use to publish the npmjs packages for NPM Trusted publishing.

Please do not set to automeerge. I will merge once I finished the setup fpr all packages on npm.

Note: For the future we need to publish new packages once manually in order to configure them for trusted publishing.