Do not import node:crypto in @matter/general
This worked because typescript was previously too lenient on types.  tsgo is stricter so we can't use node types without making them all available which is not desirable.  So instead define the subset of the Node crypto interface we use.  This is probably better anyway because it makes us more readily compatible with partial implementations from a type perspective.