Scoped Storages and FileDesignator handling
For transfers the "File designator" is basically the agreed name on the wire, but this might not match to  where wen want to store the data. This commit allows to separate these things and use one "name" for the transfer and a defined location as target or source.

It also adds flexible handling to BdxProtocol to support parallel receiving of different data (aka "ota" or "log") bia bdx and handles the direction into storages defined for the use case.